+++ src/main/java/com/palindrome/message/palindromemessageservice/repository/MessageRepository.java
@@ -3,5 +3,5 @@ package com.palindrome.message.palindromemessageservice.repository;
 import com.palindrome.message.palindromemessageservice.model.Message;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MessageRepository extends CrudRepository<Message, Integer> {
+public interface MessageRepository extends CrudRepository<Message, Long> {
 }
